First attempt at stitching
