+++ Cktgen/cktgen/cktgen/cktgen.py
@@ -1023,176 +1023,7 @@ def removeDuplicates( data):
 
     return terminals
 
-
-def consume_results(args,tech):
-    with open( 'out/' + args.block_name + '.lgf', 'rt') as fp:  
-      netl = parse_lgf( fp)
-
-    placer_results = None  
-    if args.placer_json != "":
-      with open( args.placer_json, 'rt') as fp:  
-        placer_results = json.load( fp)
-
-        
-    terminals = []
-    if placer_results is not None:
-      if args.no_interface:
-        globalScale = transformation.Transformation( 0, 0, 1, 1)
-      else:        
-        globalScale = transformation.Transformation( 0, 0, tech.halfXADTGrid*tech.pitchPoly, tech.halfYADTGrid*tech.pitchDG)
-
-      leaves_map = { leaf['template_name'] : leaf for leaf in placer_results['leaves']}
-
-      for inst in placer_results['instances']:
-        leaf = leaves_map[inst['template_name']]
-        tr = inst['transformation']
-        trans = transformation.Transformation( tr['oX'], tr['oY'], tr['sX'], tr['sY'])
-        r = globalScale.hitRect( trans.hitRect( Rect( *leaf['bbox'])).canonical())
-
-        nm = placer_results['nm'] + '/' + inst['instance_name'] + ':' + inst['template_name']
-        terminals.append( { "netName" : nm, "layer" : "cellarea", "rect" : r.toList()})
-      
-    netl.write_input_file( netl.nm + "_xxx.txt")
-
-    netl.dumpGR( tech, "INPUT/" + args.block_name + "_dr_globalrouting.json", cell_instances=terminals, no_grid=args.small)
-
-    if args.no_interface:
-      return
-
-    leaf = {}
-
-    design_name = netl.nm
-    if args.source != "":
-      design_name = args.source
-
-    leaf['template_name'] = design_name
-
-#
-# A lot to do here. This should be moved to the technology file
-#
-#
-# First assume there is only one metal template per layer
-# And only one wire width and space
-#
-    layer2MetalTemplate = {}
-    for obj in tech.metalTemplates:
-      assert obj.layer not in layer2MetalTemplate
-      layer2MetalTemplate[obj.layer] = obj
-
-    def pgd_pitch(mt):
-      assert len(mt.widths) == 2 and len(mt.spaces) == 1
-      assert mt.widths[0] == mt.widths[1]
-      return mt.widths[0] + mt.spaces[0]
-
-    def pgd_width(mt):
-      assert len(mt.widths) == 2 and len(mt.spaces) == 1
-      assert mt.widths[0] == mt.widths[1]
-      return mt.widths[0]
-
-    def ogd_pitch(mt):
-      assert len(mt.stops) == 1
-      return mt.stops[0]
-
-    def ogd_offset(mt):
-      assert len(mt.stops) == 1
-      return mt.stop_offset
-
-
-#
-# Use metal1 and metal2 for bbox grid
-#
-    shrinkX = pgd_pitch(layer2MetalTemplate['metal1'])
-    shrinkY = pgd_pitch(layer2MetalTemplate['metal2'])
-    
-    bbox = netl.bbox
-    assert bbox.llx == 0
-    assert bbox.lly == 0
-    assert bbox.urx % shrinkX == 0
-    assert bbox.ury % shrinkY == 0
-
-    leaf['bbox'] = [ bbox.llx // shrinkX, bbox.lly // shrinkY, bbox.urx // shrinkX, bbox.ury // shrinkY]
-
-    leaf['terminals'] = []
-    leaf['layout'] = []
-
-    p = re.compile('^MTI_.*|^.*_gr$')
-
-#
-# Need to do this first since via enclosure don't necessary land on the stopping point grid
-#
-    layout = []
-    for (_,wire) in netl.wire_cache.items():
-      if p.match(wire.netName): continue
-      layout.append( {
-        "net_name": wire.netName,
-        "layer": wire.layer,
-        "rect": wire.rect.toList()
-      })
-    layout = removeDuplicates(layout)
-
-    for obj in layout:
-      netName = obj['net_name']
-      rect = Rect( *obj['rect'])
-      layer = obj['layer']
-
-      if p.match(netName): continue
-
-      if layer in ["metal3","metal5","metal7"]:
-        mt = layer2MetalTemplate[layer]
-        halfWidth = pgd_width(mt) // 2
-        pgdPitch = pgd_pitch(mt)
-        ogdPitch = ogd_pitch(mt)
-        ogdOffset = ogd_offset(mt)
-
-        assert (rect.llx+halfWidth) % pgdPitch == 0
-        assert rect.lly % ogdPitch == ogdOffset, (rect.lly, rect.lly % ogdPitch)
-        assert (rect.urx-halfWidth) % pgdPitch == 0
-        assert rect.ury % ogdPitch == ogdOffset, (rect.ury, rect.ury % ogdPitch)
-
-        cx = (rect.urx + rect.llx) // (2*pgdPitch)
-        # shrink to abstract grid
-        y0 = (rect.lly + ogdOffset) // ogdPitch
-        y1 = (rect.ury - ogdOffset) // ogdPitch
-
-        leaf['terminals'].append({
-          "net_name": netName,
-          "layer": layer,
-          "rect": [ cx, y0, cx, y1]
-        })
-
-      if layer in ["metal2","metal4","metal6"]:
-        mt = layer2MetalTemplate[layer]
-        halfWidth = pgd_width(mt) // 2
-        pgdPitch = pgd_pitch(mt)
-        ogdPitch = ogd_pitch(mt)
-        ogdOffset = ogd_offset(mt)
-
-        assert (rect.lly+halfWidth) % pgdPitch == 0
-        assert rect.llx % ogdPitch == ogdOffset, (rect.llx, rect.llx % ogdPitch)
-        assert (rect.ury-halfWidth) % pgdPitch == 0
-        assert rect.urx % ogdPitch == ogdOffset, (rect.urx, rect.urx % ogdPitch)
-
-        cy = (rect.ury + rect.lly) // (2*pgdPitch)
-        # shrink to abstract grid
-        x0 = (rect.llx + ogdOffset) // ogdPitch
-        x1 = (rect.urx - ogdOffset) // ogdPitch
-
-        leaf['terminals'].append({
-          "net_name": netName,
-          "layer": layer,
-          "rect": [ x0, cy, x1, cy]
-        })
-
-    leaf['terminals'] = removeDuplicates(leaf['terminals'])
-    leaf['layout'] = layout
-
-    interface_fn = "INPUT/interface.json"
-    if args.source != '':
-      interface_fn = "INPUT/" + args.source + "_interface.json"
-
-    with open( interface_fn, "wt") as fp:
-      fp.write( json.dumps( { "leaves": [ leaf ]}, indent=2) + "\n")
-
+from .consume_results import consume_results
 
 def parse_args( command_line_args=None):
   parser = argparse.ArgumentParser( description="Generates input files for amsr (Analog router)")
