+++ Viewer/index.html
@@ -91,7 +91,10 @@
       { layer: "metal4", master: "M4" },
       { layer: "metal5", master: "M5" },
       { layer: "metal6", master: "M6" },
-      { layer: "boundary", master: "diearea" }
+      { layer: "boundary", master: "diearea" },
+      { layer: "diffcon", master: "Diffcon" },
+      { layer: "polycon", master: "Polycon" },
+      { layer: "poly", master: "Poly" }
     ];
 
     var metalStack = [
